1、 去掉业务配置新增时 业务树 拖拽功能 修复 出现多个业务根节点bug 2、 【全局配置】删除跟节点后提示成功，实际未生效 3、【全局配置】删除api和业务时，建议给出确认弹框以免误删

+++ src/views/configuration/visitCall/Detail.vue
@@ -128,16 +128,20 @@ export default {
      this.currentPars = this.currentRow[ this.editableTabsValue].parseParams
     },
     deleteApis() {
-      console.log(this.editableTabsValue)
-
-      let currentApi = this.currentRow[this.editableTabsValue]
-      deleteApi(currentApi.id).then(res => {
-        this.currentRow.splice(+this.editableTabsValue,1)
-        console.log(this.currentRow)
-        this.editableTabsValue= '0'
-        this.$message({
-          type: 'success',
-          message: '删除成功'
+      this.$confirm('此操作将删除当前api, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        let currentApi = this.currentRow[this.editableTabsValue]
+        deleteApi(currentApi.id).then(res => {
+          this.currentRow.splice(+this.editableTabsValue,1)
+          console.log(this.currentRow)
+          this.editableTabsValue= '0'
+          this.$message({
+            type: 'success',
+            message: '删除成功'
+          })
         })
       })
     }
